+++ spec/whois/record/parser/responses/whois.srs.net.nz/status_available_spec.rb
@@ -57,4 +57,9 @@ describe Whois::Record::Parser::WhoisSrsNetNz, "status_available.expected" do
       @parser.nameservers.should == []
     end
   end
+  describe "#response_throttled?" do
+    it do
+      @parser.response_throttled?.should == false
+    end
+  end
 end