+++ spec/whois/record/parser/responses/whois.srs.net.nz/status_registered_spec.rb
@@ -68,4 +68,9 @@ describe Whois::Record::Parser::WhoisSrsNetNz, "status_registered.expected" do
       @parser.nameservers[3].name.should == "ns4.google.com"
     end
   end
+  describe "#response_throttled?" do
+    it do
+      @parser.response_throttled?.should == false
+    end
+  end
 end